requirements.txt
# Python dependencies for TSCM Signature Analysis in SLE

# Core data analysis
scanpy>=1.9.0
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
anndata>=0.8.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
Pillow>=9.0.0

# Machine learning and statistics
scikit-learn>=1.1.0
statsmodels>=0.13.0

# Pathway enrichment
gseapy>=1.0.0

# Document processing (for manuscript preparation)
python-docx>=0.8.11

# Utilities
tqdm>=4.64.0

